fix otp

+++ src/main/java/com/doan/VitaStore/repository/PasswordResetTokenRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetTokenEntity, Long> {
+public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetTokenEntity, Integer> {
     Optional<PasswordResetTokenEntity> findByEmailAndOtpAndUsedFalse(String email, String otp);
     Optional<PasswordResetTokenEntity> findByEmailAndTokenAndUsedFalse(String email, String token);
     Optional<PasswordResetTokenEntity> findTopByEmailOrderByOtpExpiryDesc(String email);
